update13

+++ schema.sql
@@ -15,7 +15,7 @@ CREATE TABLE Customers (
     customer_id INT AUTO_INCREMENT PRIMARY KEY,
     first_name VARCHAR(255) NOT NULL,
     last_name VARCHAR(255) NOT NULL,
-    rating INT,
+    rating FLOAT,
     company_id INT,
     FOREIGN KEY (company_id) REFERENCES Companies(company_id)
 );
